# course-service/src/main/java/com/sma/course/dto/QuizQuestion.java
package com.sma.course.dto;

import java.util.List;

public record QuizQuestion(
        String questionId,
        String stem,
        List<String> options,
        String answer,
        String explanation
) {
}